refactor(App): relocate the App component to pages
