+++ src/pages/App/App.jsx
@@ -3,39 +3,39 @@ import {
   Navigate, Route, Routes, useNavigate,
 } from 'react-router-dom';
 import './App.scss';
-import ProtectedRoute from '../ProtectedRoute';
+import ProtectedRoute from '../../components/ProtectedRoute';
 import Header from '../../modules/Header/Header';
 import Footer from '../../modules/Footer/Footer';
-import MainPage from '../../pages/MainPage/MainPage';
-import MapPage from '../../pages/MapPage/MapPage';
-import SheltersListPage from '../../pages/SheltersListPage/SheltersListPage';
-import * as sheltersListModules from '../../pages/SheltersListPage/modules';
-import ShelterPage from '../../pages/ShelterPage/ShelterPage';
-import * as shelterModules from '../../pages/ShelterPage/modules';
-import PetPage from '../../pages/PetPage/PetPage';
-import AddShelterPage from '../../pages/AddShelterPage/AddShelterPage';
-import PapersPage from '../../pages/PapersPage/PapersPage';
-import PaperPage from '../../pages/PaperPage/PaperPage';
-import NewsPage from '../../pages/NewsPage/NewsPage';
-import NewPage from '../../pages/NewPage/NewPage';
-import LoginPage from '../../pages/LoginPage/LoginPage';
-import RegisterPage from '../../pages/RegisterPage/RegisterPage';
-import NotFoundPage from '../../pages/NotFoundPage/NotFoundPage';
-import PasswordRecovery from '../../pages/PasswordRecovery/PasswordRecovery';
-import NewPassword from '../../pages/NewPassword/NewPassword';
-import SignUpConfirm from '../../pages/SignUpConfirm/SignUpConfirm';
-import ProfilePage from '../../pages/ProfilePage/ProfilePage';
+import MainPage from '../MainPage/MainPage';
+import MapPage from '../MapPage/MapPage';
+import SheltersListPage from '../SheltersListPage/SheltersListPage';
+import * as sheltersListModules from '../SheltersListPage/modules';
+import ShelterPage from '../ShelterPage/ShelterPage';
+import * as shelterModules from '../ShelterPage/modules';
+import PetPage from '../PetPage/PetPage';
+import AddShelterPage from '../AddShelterPage/AddShelterPage';
+import PapersPage from '../PapersPage/PapersPage';
+import PaperPage from '../PaperPage/PaperPage';
+import NewsPage from '../NewsPage/NewsPage';
+import NewPage from '../NewPage/NewPage';
+import LoginPage from '../LoginPage/LoginPage';
+import RegisterPage from '../RegisterPage/RegisterPage';
+import NotFoundPage from '../NotFoundPage/NotFoundPage';
+import PasswordRecovery from '../PasswordRecovery/PasswordRecovery';
+import NewPassword from '../NewPassword/NewPassword';
+import SignUpConfirm from '../SignUpConfirm/SignUpConfirm';
+import ProfilePage from '../ProfilePage/ProfilePage';
 import CurrentUserContext from '../../contexts/CurrentUserContext';
 import * as auth from './api/auth';
 import imageSuccess from '../../images/icons/ic_success.svg';
 import imageError from '../../images/icons/ic_error.svg';
-import InfoTooltip from '../InfoTooltip/InfoTooltip';
+import InfoTooltip from '../../components/InfoTooltip/InfoTooltip';
 import * as userApi from './api/userApi';
 import { register } from './api/auth';
-import EditProfilePage from '../../pages/EditProfilePage/EditProfilePage';
-import SignOutPage from '../../pages/SignOutPage/SignOutPage';
-import ChangePasswordPage from '../../pages/ChangePasswordPage/ChangePasswordPage';
-import ActivateUserPage from '../../pages/ActivateUserPage/ActivateUserPage';
+import EditProfilePage from '../EditProfilePage/EditProfilePage';
+import SignOutPage from '../SignOutPage/SignOutPage';
+import ChangePasswordPage from '../ChangePasswordPage/ChangePasswordPage';
+import ActivateUserPage from '../ActivateUserPage/ActivateUserPage';
 
 const App = () => {
   const navigate = useNavigate();
