+++ src/pages/App/styles/page.scss
