+++ src/pages/App/styles/standart-font.scss
